pyproject.toml
[tool.poetry]
name = "spotdl"
version = "4.0.0-rc.1"
description = "Download your Spotify playlists and songs along with album art and metadata"
license = "MIT"
authors = ["spotDL Team <spotdladmins@googlegroups.com>"]
maintainers = ["xnetcat <xnetcat.dev@gmail.com>"]
readme = "README.md"
repository = "https://github.com/spotDL/spotify-downloader.git"
homepage = "https://github.com/spotDL/spotify-downloader/"
documentation = "https://spotdl.rtfd.io/en/latest/"
keywords = ["spotify", "downloader", "spotdl", "music"]
classifiers = [
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Multimedia",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Utilities",
]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"

spotipy = "^2.19.0"
ytmusicapi = "^0.22.0"
pytube = "^12.1.0"
yt-dlp = "^2022.5.18"
mutagen = "^1.45.1"
rich = "^12.4.4"
beautifulsoup4 = "^4.10.0"
requests = "^2.27.1"
rapidfuzz = "^2.0.11"
python-slugify = "^6.1.1"
uvicorn = "^0.17.1"
pydantic = "^1.9.0"
fastapi = "^0.78.0"
nest_asyncio = "^1.5.4"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-mock = "^3.7.0"
pytest-vcr = "^1.0.2"
pyfakefs = "^4.5.4"
pytest-cov = "^3.0.0"
pytest-subprocess = "^1.4.0"
pytest-asyncio = "^0.18.3"
mypy = "^0.961"
pylint = "^2.13.9"
black = "^22.1.0"
mdformat-gfm = "^0.3.5"
types-orjson = "^3.6.2"
types-python-slugify = "^5.0.3"
types-requests = "^2.27.16"
types-setuptools = "^57.4.11"
types-toml = "^0.10.4"
types-ujson = "^5.3.0"
pyinstaller = "^5.1"
mkdocs = "^1.3.0"
mkdocs-material = "^8.3.3"
mkdocstrings = "^0.19.0"
mkdocstrings-python = "^0.7.0"
pymdown-extensions = "^9.5"
mkdocs-gen-files = "^0.3.4"
mkdocs-literate-nav = "^0.4.1"
mkdocs-section-index = "^0.3.4"

[tool.poetry.scripts]
spotdl = "spotdl:console_entry_point"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.format]
limit-inference-results = 0
fail-under = 9

[mypy]
ignore_missing_imports = true
